requirements.txt
langchain
langchain-community
langchain-openai
faiss-cpu
streamlit
tiktoken
PyPDF2
python-dotenv
